Use queue to buffer native evo watcher values
With the iterator used until now it could not be easily
determined when all the watched data were read (e.g., in list comprehensions).
Closing the generator could signal the watcher is done, but it caused
the unread data to be lost.

This change uses queue to buffer the data instead, making
reading independent of the actual watcher state.